Need to tweak config to trigger data timeout more frequently

+++ src/main/scala/com/gtan/repox/GetAsyncHandler.scala
@@ -20,25 +20,27 @@ class GetAsyncHandler(val uri: String,
                       val tempFilePath: Option[String]) extends AsyncHandler[Unit] with LazyLogging {
 
   private var tempFileOs: FileOutputStream = null
-  private var tempFileChannel: FileChannel = null
   var tempFile: File = null
   private var contentType: String = null
 
   @volatile private var canceled = false
 
-  def cancel() {
+  def cancel(deleteTempFile: Boolean = true) {
     canceled = true
-    cleanup()
+    cleanup(deleteTempFile)
   }
 
   override def onThrowable(t: Throwable): Unit = {
-    worker ! AsyncHandlerThrows(t)
+    if (!canceled)
+      worker ! AsyncHandlerThrows(t)
   }
 
   override def onCompleted(): Unit = {
     if (!canceled) {
-      if (tempFileOs != null)
+      if (tempFileOs != null) {
+        logger.debug(s"onCompleted: tempFile.length = ${tempFile.length()}")
         tempFileOs.close()
+      }
       if (tempFile != null) {
         // completed before parent notify PeerChosen or self cancel
         master.!(Completed(tempFile.toPath, repo))(worker)
@@ -48,8 +50,8 @@ class GetAsyncHandler(val uri: String,
 
   override def onBodyPartReceived(bodyPart: HttpResponseBodyPart): STATE = {
     if (!canceled) {
-      val written = tempFileChannel.write(bodyPart.getBodyByteBuffer)
-      worker ! PartialDataReceived(written)
+      bodyPart.writeTo(tempFileOs)
+      worker ! PartialDataReceived(bodyPart.length())
       STATE.CONTINUE
     } else {
       cleanup()
@@ -58,15 +60,16 @@ class GetAsyncHandler(val uri: String,
   }
 
   override def onStatusReceived(responseStatus: HttpResponseStatus): STATE = {
-    if (responseStatus.getStatusCode == StatusCodes.NOT_FOUND) {
+    val statusCode: Int = responseStatus.getStatusCode
+    if (statusCode == StatusCodes.NOT_FOUND) {
       Repox.head404Cache ! NotFound(uri, repo)
     }
     if (canceled) {
       cleanup()
       STATE.ABORT
     } else {
-      if (responseStatus.getStatusCode > 200 | responseStatus.getStatusCode >= 300) {
-        logger.debug(s"Get ${repo.absolute(uri)} ${responseStatus.getStatusCode}")
+      if (statusCode < 200 || statusCode >= 400) {
+        logger.debug(s"Get ${repo.absolute(uri)} $statusCode")
         master.!(UnsuccessResponseStatus(responseStatus))(worker)
         cleanup()
         STATE.ABORT
@@ -85,14 +88,10 @@ class GetAsyncHandler(val uri: String,
           logger.debug("Lantern interrupted. Resync data.")
           tempFileOs.close()
           tempFileOs = new FileOutputStream(tempFile)
-          tempFileChannel = tempFileOs.getChannel
-          tempFileChannel.position(tempFileChannel.size())
           worker ! HeadersGot(headers)
         } else {
-          tempFile = newOrReuseTempFile(tempFilePath)
-          tempFileOs = new FileOutputStream(tempFile)
-          tempFileChannel = tempFileOs.getChannel
-          tempFileChannel.position(tempFileChannel.size())
+          tempFile = newOrReuseTempFile()
+          tempFileOs = new FileOutputStream(tempFile, true)
           worker ! HeadersGot(headers)
           master.!(HeadersGot(headers))(worker)
         }
@@ -108,25 +107,22 @@ class GetAsyncHandler(val uri: String,
     }
   }
 
-  def cleanup(): Unit = {
+  def cleanup(deleteTempFile: Boolean = true): Unit = {
     if (tempFileOs != null) {
       tempFileOs.close()
     }
-    if (tempFile != null) {
+    if (tempFile != null && deleteTempFile) {
       tempFile.delete()
     }
-    worker ! PoisonPill
   }
 
-  private def newOrReuseTempFile(path: Option[String]): File = {
-    path match {
-      case None =>
-        val parent = Config.storagePath.resolve("temp").toFile
-        parent.mkdirs()
-        File.createTempFile("repox", ".tmp", parent)
-      case Some(file) =>
-        new File(file)
-    }
+  private def newOrReuseTempFile(): File = tempFilePath match {
+    case None =>
+      val parent = Config.storagePath.resolve("temp").toFile
+      parent.mkdirs()
+      File.createTempFile("repox", ".tmp", parent)
+    case Some(file) =>
+      new File(file)
   }
 
 }
